// src/modules/garmentsProduction/reports/finishing/finish-fabric-return-cutting-floor-to-store-report/finish-fabric-return-cutting-floor-to-store-report-type.ts
export interface IFinishFabricReturnCuttingFloorToStoreReport {
  BUYER_ID: number;
  STYLE_ID: number;
  PO_ID: number;
  FABRIC_ID: number;
  BUYER_NAME: string;
  STYLE_NO: string;
  PO_NO: string;
  FABRIC_NAME: string;
  REQ_QTY: number;
  REQSITION_QTY: number;
  RCV_QTY: number;
  RETURN_QTY: number;
}